made fixes on handling.
used slick-codegen 3.1.0 for generating Tables.scala
then I replaced all "0.00" to BigDecimal("0.00") as this was a Bug in the code generator

Used futures for getting profile.

Issue here is the json transformation. Due to the restrictions in case class (22 parameters) the Members table was converted
to a trait and def. Thereby would need custom Json format writers.